+++ comprl/CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   config file location instead of to the working directory.
 - Matchmaking now samples from all candidates with quality above the threshold instead
   of using the first in the list.
+- Do not update player ratings for games with disconnect.
+- Add option to limit number of games run in parallel.
 
 ## Removed
 - The `Agent.event` decorator has been removed.  Instead of using it, create